TOOLS: release gm2xpkg ver 0.10 (cfg 0.9) (#109)
* use tmp path for default $RELEASEDIR $OPKG_ASSETSDIR

* allow MANUAL to be another type e.g. `Unicode text`

* separate gm2xpkg & pkg.cfg VER envar

use $PKGCFG_VER

* mkdir selectordir if present var

minor shellcheck warn fix

* use `postinst` instead for mkdir $SELDIR

+++ tools/gm2xpkg.sh
@@ -1,6 +1,7 @@
 #!/bin/bash
 
-VER=0.9
+VER=0.10
+PKGCFG_VER=0.9
 MIYOOCFW_VER=2.0.0
 # Help & About info
 help_func() {
@@ -16,7 +17,7 @@ help_func() {
 	 \t -z, --zip       generate ZIP archive
 	 \t -p, --pkg       generate ./package
 	 \t -c, --clean     remove ./package ./opkg_assets ./<target_name>.ipk ./<target_name>.zip ./<link_name>lnk
-	 \t -g, --gencfg    generate standard config \"pkg.cfg\" file in PWD
+	 \t -g, --gencfg    generate standard config \"pkg.cfg\" v${PKGCFG_VER} file in PWD
 	 \t -f, --force     force execution even without present target's binary or incompatible .cfg version
 	 \t -q, --quiet     don't output messages to stdout
 	 Instructions:
@@ -44,7 +45,7 @@ pkg_config_func() {
 	echo "# NOTES:
 ## All variable values should be enclosed within double quotes: \"<value>\"
 ## CONFIGURATION FILE for \`gm2xpkg\` script version:
-PKGVER=\"${VER}\"
+PKGVER=\"${PKGCFG_VER}\"
 
 # EXEC commands (set to \"1\" anyone for desired outcome), you can instead use [OPTIONS] of \`gm2xpkg\`:
 PACKAGE=\"${PACKAGE}\"
@@ -199,7 +200,7 @@ do
 						echo "NO, exiting..."
 						;;
 					[Uu]*)
-						echo "Update, upgrading present ${PKGCFG} file with new config version - ${VER}"
+						echo "Update, upgrading present ${PKGCFG} file with new config version - ${PKGCFG_VER}"
 						UPDATE_PKGCFG=yes
 						pkg_config_func
 						;;
@@ -248,8 +249,8 @@ if test -f "${PKGCFG}"; then
 		echo "<<<EOF>>>"
 	fi
 	source "${PKGCFG}"
-	if test "${VER}" != "${PKGVER}" ; then
-		echo -e "ERROR: GM2X PACKAGER version ${VER} doesn't match CONFIGURATION FILE version ${PKGVER}\n\n\tPlease update your ${PKGCFG} config file, use [--gencfg] option\n"
+	if test "${PKGCFG_VER}" != "${PKGVER}" ; then
+		echo -e "ERROR: GM2X PACKAGER CFG version ${PKGCFG_VER} doesn't match CONFIGURATION FILE version ${PKGVER}\n\n\tPlease update your ${PKGCFG} config file, use [--gencfg] option\n"
 		if test "x${FORCE}" == "xyes"; then
 			echo "WARNING: Force-mode active, running regardless deprecated/incompatible ${PKGCFG}."
 		else
@@ -307,9 +308,9 @@ fi
 
 ## Generic
 HOMEPATH=${HOMEPATH:="/mnt"}
-RELEASEDIR=${RELEASEDIR:=package}
+RELEASEDIR=${RELEASEDIR:=$(mktemp -d)}
 ASSETSDIR=${ASSETSDIR:=assets}
-OPKG_ASSETSDIR=${OPKG_ASSETSDIR:=opkg_assets}
+OPKG_ASSETSDIR=${OPKG_ASSETSDIR:=$(mktemp -d)}
 LINK=${LINK:=$TARGET.lnk}
 ALIASES=${ALIASES:=aliases.txt}
 MANUAL=${MANUAL:=$TARGET.man.txt}
@@ -465,12 +466,15 @@ fi
 if test $PACKAGE -eq 1 >/dev/null 2>&1 || test $ZIP -eq 1 >/dev/null 2>&1 || test $IPK -eq 1 >/dev/null 2>&1; then
 	TARGET_INSTALL_DIR=$RELEASEDIR/$DESTDIR/$TARGET_DIR
 	# Create ./package
-	rm -rf $RELEASEDIR
-	mkdir -p $RELEASEDIR
-	# mkdir -p $ASSETSDIR
-	mkdir -p $OPKG_ASSETSDIR
+	rm -rf "${RELEASEDIR:?}"/*
+	#mkdir -p $RELEASEDIR
+	#mkdir -p $ASSETSDIR
+	#mkdir -p $OPKG_ASSETSDIR
 	mkdir -p $TARGET_INSTALL_DIR
 	mkdir -p $RELEASEDIR/gmenu2x/sections/$SECTION
+	# TODO: wait for patch in opkg or fix symlink deletion when unistalling ipk
+	#test -n "$SELDIR" \
+	#&& mkdir -p ${RELEASEDIR}${SELDIR}
 	if test "x${FORCE}" != "xyes" || test "x${TARGET_EXIST}" == "xyes"; then
 		cp $TARGET_PATH $TARGET_INSTALL_DIR/
 	else
@@ -509,7 +513,7 @@ if test $PACKAGE -eq 1 >/dev/null 2>&1 || test $ZIP -eq 1 >/dev/null 2>&1 || tes
 	if test -e $MANUAL; then
 		if file $MANUAL | grep -q "PNG image"; then
 			MANUAL_EXT=".man.png"
-		elif file $MANUAL | grep -q "ASCII text"; then
+		elif file $MANUAL | grep -q "text"; then
 			MANUAL_EXT=".man.txt"
 		else
 			MANUAL_EXT=""
@@ -529,9 +533,9 @@ if test $PACKAGE -eq 1 >/dev/null 2>&1 || test $ZIP -eq 1 >/dev/null 2>&1 || tes
 		done
 	fi
 	if test -d $RELEASEDIR/gmenu2x && test -d $TARGET_INSTALL_DIR; then
-	 	test $PACKAGE -eq 1 && echo "Done packaging ./$RELEASEDIR/ data" || echo "Ready to use ./$RELEASEDIR/ data for deaper packaging"
+	 	test $PACKAGE -eq 1 && echo "Done packaging $RELEASEDIR data" || echo "Ready to use $RELEASEDIR/ data for deaper packaging"
 	else
-		echo "WARNING: Upss smth went wrong and I couldn't locate auto-gen data in ./$RELEASEDIR/"
+		echo "WARNING: Upss smth went wrong and I couldn't locate auto-gen data in $RELEASEDIR/"
 	fi
 	
 	# Create ./package/<target_version>.zip
@@ -546,14 +550,20 @@ if test $PACKAGE -eq 1 >/dev/null 2>&1 || test $ZIP -eq 1 >/dev/null 2>&1 || tes
 	# Create ./package/<target>.ipk
 	if test $IPK -eq 1 >/dev/null 2>&1; then
 		# rm -rf $RELEASEDIR/*.ipk $RELEASEDIR/*.zip
-		mkdir -p .$HOMEPATH
-		cp -r $RELEASEDIR/* .$HOMEPATH && mv .$HOMEPATH $RELEASEDIR/
+		mkdir -p $RELEASEDIR/$HOMEPATH
+		for f in $RELEASEDIR/*; do test "$f" != "${RELEASEDIR}${HOMEPATH}" && mv "$f" $RELEASEDIR/$HOMEPATH; done
 		mkdir -p $RELEASEDIR/data
 		mv $RELEASEDIR$HOMEPATH $RELEASEDIR/data/
 		if ! { test -d ${OPKG_ASSETSDIR}/CONTROL && test -f ${OPKG_ASSETSDIR}/CONTROL/preinst && test -f ${OPKG_ASSETSDIR}/CONTROL/postinst && test -f ${OPKG_ASSETSDIR}/CONTROL/control; }; then
 			mkdir -p $OPKG_ASSETSDIR/CONTROL
+			#preinst script
 			echo -e "#!/bin/sh\nsync; echo 'Installing new ${TARGET}..'; rm /var/lib/opkg/info/${TARGET}.list; exit 0" > $OPKG_ASSETSDIR/CONTROL/preinst
-			echo -e "#!/bin/sh\nsync; echo 'Installation finished.'; echo 'Restarting ${TARGET}..'; sleep 1; killall ${TARGET}; exit 0" > $OPKG_ASSETSDIR/CONTROL/postinst
+			#postinst script
+			echo -e "#!/bin/sh\nsync;" >> $OPKG_ASSETSDIR/CONTROL/postinst
+			test -n "$SELDIR" \
+			 && echo -e "! test -d ${SELDIR} && mkdir -p ${SELDIR} && echo 'Created ${SELDIR} dir for Selector.';" >> $OPKG_ASSETSDIR/CONTROL/postinst
+			echo -e "echo 'Installation finished.'; echo 'Restarting ${TARGET}..'; sleep 1; killall ${TARGET}; exit 0" >> $OPKG_ASSETSDIR/CONTROL/postinst
+			#control file
 			echo -e $CONTROL > $OPKG_ASSETSDIR/CONTROL/control
 		else
 			OPKG_ASSETSDIR_CUSTOM="yes"
